Userstory1 60% finished
One can now access the page for orders and put a random bottom and topping of their liking.
The database has been updated so that it now stores the toppings and the Bottoms. A user table has been made for the login of the users, whether its an employee or customer